refactor(log): common log namespace (#419)
* refactor(log): common log namespace

* refactor(log): localizedLogger extend psr Abstract logger

* refactor(log): localizedLoggerFactory plus contracts

ems.logger will be defined in the coreBundle (only used there)
