+++ src/Common/Log/DoctrineHandler.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace EMS\CommonBundle\Helper\Logger;
+namespace EMS\CommonBundle\Common\Log;
 
 use Doctrine\DBAL\Types\Types;
 use Doctrine\ORM\EntityManagerInterface;
