+++ src/Contracts/Log/LocalizedLoggerInterface.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace EMS\CommonBundle\Contracts\Log;
+
+use Psr\Log\LoggerInterface;
+
+interface LocalizedLoggerInterface extends LoggerInterface
+{
+}
